Делаю то же самое в level2.java

+++ src/com/example/smartalarm2/Level2.java
@@ -33,7 +33,7 @@ Random Sklizenka = new Random();
 		int resultresult2 = result2-result;
 		int c = randomik.nextInt(2);
 		int plusminus = Sklizenka.nextInt(2);
-		double endResult = 0 ;
+		int endResult = 0 ;
         Button button1aa = (Button)findViewById(R.id.button1a);
         Button button2aa = (Button)findViewById(R.id.button2a);
         Button button3aa = (Button)findViewById(R.id.button3a);
@@ -47,8 +47,8 @@ Random Sklizenka = new Random();
 			endResult = resultresult2;
 			textView1.setText(result2+"-"+result+"=?");
 		}
-		double wResult = endResult +(Math.pow(10,randomik.nextInt(2)))*(randomik.nextInt(8)+1); 
-		double wResult2 = endResult +(Math.pow(10,randomik.nextInt(2)))*(randomik.nextInt(8)+1); 
+		int wResult = (int)endResult +(Math.pow(10,randomik.nextInt(2)))*(randomik.nextInt(8)+1); 
+		int wResult2 = (int)endResult +(Math.pow(10,randomik.nextInt(2)))*(randomik.nextInt(8)+1); 
 		
 		 if(c==0){
 			 button1aa.setText(""+endResult+"");
